+++ src/piclassifier/__init__.py
@@ -1,0 +1,4 @@
+def pi_classify():
+    from piclassifier.piclassify import main
+
+    main()
