+++ src/piclassifier/processor.py
@@ -19,7 +19,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 from abc import ABC, abstractmethod
 
-from service import SnapshotService
+from .service import SnapshotService
 
 
 class Processor(ABC):
